move SpoonMetaModel from test to main (runtime)
