+++ src/main/java/spoon/metamodel/MMTypeKind.java
@@ -14,7 +14,7 @@
  * The fact that you are presently reading this means that you have had
  * knowledge of the CeCILL-C license and that you accept its terms.
  */
-package spoon.test.metamodel;
+package spoon.metamodel;
 
 /**
  * Represents type of Spoon model class
