+++ src/test/java/spoon/test/api/MetamodelTest.java
@@ -6,6 +6,10 @@ import spoon.Launcher;
 import spoon.Metamodel;
 import spoon.SpoonAPI;
 import spoon.SpoonException;
+import spoon.metamodel.MMMethodKind;
+import spoon.metamodel.MetamodelConcept;
+import spoon.metamodel.MetamodelProperty;
+import spoon.metamodel.SpoonMetaModel;
 import spoon.reflect.annotations.MetamodelPropertyField;
 import spoon.reflect.annotations.PropertyGetter;
 import spoon.reflect.annotations.PropertySetter;
@@ -33,10 +37,6 @@ import spoon.reflect.visitor.filter.AnnotationFilter;
 import spoon.reflect.visitor.filter.SuperInheritanceHierarchyFunction;
 import spoon.reflect.visitor.filter.TypeFilter;
 import spoon.template.Parameter;
-import spoon.test.metamodel.MMMethodKind;
-import spoon.test.metamodel.MetamodelConcept;
-import spoon.test.metamodel.MetamodelProperty;
-import spoon.test.metamodel.SpoonMetaModel;
 
 import java.io.File;
 import java.util.ArrayList;
