+++ src/test/java/spoon/test/replace/ReplaceParametrizedTest.java
@@ -4,6 +4,10 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 import spoon.SpoonException;
+import spoon.metamodel.MMTypeKind;
+import spoon.metamodel.MetamodelConcept;
+import spoon.metamodel.MetamodelProperty;
+import spoon.metamodel.SpoonMetaModel;
 import spoon.reflect.code.CtBlock;
 import spoon.reflect.code.CtFieldAccess;
 import spoon.reflect.code.CtStatement;
@@ -19,10 +23,6 @@ import spoon.reflect.reference.CtTypeReference;
 import spoon.reflect.visitor.CtScanner;
 import spoon.reflect.visitor.CtVisitable;
 import spoon.reflect.visitor.Filter;
-import spoon.test.metamodel.MMTypeKind;
-import spoon.test.metamodel.MetamodelConcept;
-import spoon.test.metamodel.MetamodelProperty;
-import spoon.test.metamodel.SpoonMetaModel;
 
 import java.io.File;
 import java.util.ArrayList;
